Initial commit of a typing window component
Create a component where the typing test is administered. Based on a
word list (currently hard-coded), each word is mapped to an object that
tracks what the expected word is and what the user actually typed.
React state tracks various metrics such as the current index in the word
list, the current word, the expected letter, the actual typed letters,
etc.

A keydown event listener is added to the window on mount (and removed on
on unmount) via useEffect, and the callback for the event listener
handles all keypresses, updating state accordingly. The expected words
and typed words are consolidated in an array of objects and this is
mapped to Word components for rendering.