Remove useMemo hook
For hard-coded list of words for the typing test, useMemo was originally
suggested by the linter since the array itself was referenced as a
dependency in the useEffect hook that handles keydown events. This array
was removed from being a dependency in favor of the `wordsObject` object
so having this array remain defined via useMemo is now unnecessary.