Include extra and missed letters in calculations
Net WPM should consider all uncorrected errors, so wrong letters,
extra letters, and missed letters as determined at the completion of the
typing test. Errors that were corrected are inherently reflected in the
WPM calculation (raw or net) since time would be required to go back and
fix them.

Accuracy considers all errors typed as well as extra and missed letters.
This is more of a "live" accuracy, or the measure of whether the typed
letter at any given time matched what the expected letter was. For extra
letters, the error is typing letters that were not space, thus moving
the user on to the next word. Likewise for missing letters, the error is
not typing the letter and instead skipping to the next word. The user
would be penalized for all letters skipped in a given word.