Add backend fetching and modularize word generation
Add fetch calls in useEffect hooks to send GET requests to backend
for different word lists.

Split up and modularize how the `wordsObject` is created including
a function to map a new array from a plain array of strings (the words)
as well as another to randomize the selection for inclusion in the
typing test.