Fix poor spacing when rendering hidden words
renderLetters function returns null if the row of a given word is
outside the range specified in the logic. Even though null is returned,
it is passed to the div in the final render of the return overlal
function's return statement, so the div is still passed. The margin
specified in the div (tailwindcss mr-[14.4px]) was causing odd blank
spaces. Adding a condition for rendering the entire div so that it only
renders when a word would render fixes the issue.