Update cursor to only have two vertical positions
With a maximum of only 3 rows of words being rendered to the typing
window, the cursor should be on the first row only at the beginning of
the test, and as the test proceeds and the rows of words are scrolled
continuously, the cursor should stay in the middle. This corresponds to
0 and 1 respectively, which will be multiplied by the 32px that is set
in the CSS `top` property. This effectively keeps the cursor vertically
positioned as desired.