Limit keydown events if test is already completed
Ignoring keydown events after the test is complete is simple way of
preserving accuracy calcs after the test has finished, so no more
letters can be typed and determined to be correct, error, missed, extra,
etc.

Adding a single option for now which is to press 'Enter' to reset the
test.