Animate typing window with fade animation and key
React forces unmount and remount of component with a key specified that
changes, so create a new state variable for the typing window's key and
have it be changed as needed i.e. in configuration changes. This will
prevent the unsightly and jagged transitions from occuring. Incorporate
the custom fade in animation defined in the tailwind config file to
further smoothen the visual transition.