Update loading message location and add animation
Move loading message to the same line as other feedback messages such as
the one to begin typing to start the test, timer, word count, and marking
the test as complete. The loading status must be propogated to these
other messages as well so that only one message is ever visible at any
given time. Incorporating the fade in animation to all the messages also
improves the visual experience.